+++ crates/lakekeeper-storage-postgres/.env.sample
@@ -1,0 +1,2 @@
+# This is the URL that SQLx will use to make connections to the Postgres database.
+DATABASE_URL=postgresql://postgres:postgres@localhost/postgres
