+++ crates/lakekeeper/src/api/management/v1/tasks.rs
@@ -1809,675 +1809,4 @@ mod test {
             assert!(validate_schedule_request_static_checks(&req, now()).is_ok());
         }
     }
-
-    /// Postgres-backed lifecycle of the schedule endpoint.
-    ///
-    /// Covers the chain that the in-process unit tests can't:
-    ///   1. First schedule → `200 { task_id }`
-    ///   2. Second schedule → `409` with `task-id=<first>` in the body
-    ///      (proves the conflict-lookup path through
-    ///      `lookup_active_task_id` + `format_task_already_active_error`)
-    ///   3. `task/control RunNow` on the existing id → `204`
-    ///
-    /// Uses the `setup_and_registry` test helper to register a custom
-    /// `user_schedulable = true` test queue so the OSS endpoint has
-    /// something to dispatch to without enabling enterprise queues.
-    // Postgres-dependent; extracted to integration-tests as a follow-up.
-    #[cfg(any())]
-    mod schedule_lifecycle {
-        use std::sync::{Arc, LazyLock};
-
-        use iceberg::spec::{Schema, UnboundPartitionSpec};
-        use iceberg_ext::catalog::rest::CreateTableRequest;
-        use lakekeeper_storage_postgres::tests::{memory_io_profile, setup_with_registry};
-        use serde::{Deserialize, Serialize};
-        use sqlx::PgPool;
-
-        use crate::{
-            api::{
-                iceberg::v1::{
-                    DataAccess, NamespaceParameters, Prefix, tables::TablesService as _,
-                },
-                management::v1::{
-                    ApiServer,
-                    task_queue::{ScheduleTaskRequest, ScheduleTaskResponse},
-                    tasks::{ControlTaskAction, ControlTasksRequest, Service as _},
-                    warehouse::TabularDeleteProfile,
-                },
-            },
-            request_metadata::RequestMetadata,
-            server::CatalogServer,
-            service::{
-                TableId,
-                authz::AllowAllAuthorizer,
-                tasks::{
-                    QueueRegistration, QueueScope, TaskConfig, TaskData, TaskQueueName,
-                    UserScheduling, WarehouseTaskEntityId,
-                },
-            },
-        };
-
-        static TEST_QUEUE_NAME: LazyLock<TaskQueueName> =
-            LazyLock::new(|| "test_schedulable_lifecycle".into());
-        static REJECTING_QUEUE_NAME: LazyLock<TaskQueueName> =
-            LazyLock::new(|| "test_schedulable_rejecting".into());
-
-        /// Marker property the rejecting queue's eligibility check looks at.
-        /// When set to `"reject"` on a table the queue refuses to schedule.
-        const REJECTION_MARKER_PROPERTY: &str = "schedule-test.reject-me";
-
-        /// Empty payload shared by both test queues. Real queues bind their
-        /// own `TaskData`; the lifecycle test just needs the type to round-
-        /// trip through the payload validator.
-        #[derive(Clone, Debug, Default, Serialize, Deserialize)]
-        struct TestSchedulablePayload {}
-        impl TaskData for TestSchedulablePayload {}
-
-        /// Minimal user-schedulable queue for the lifecycle test. No worker
-        /// is registered (`num_workers=0`) — we never want a real run, just
-        /// the schedule/conflict/control round-trips on the catalog row.
-        #[derive(Clone, Debug, Default, Serialize, Deserialize)]
-        #[cfg_attr(feature = "open-api", derive(utoipa::ToSchema))]
-        struct TestSchedulableConfig {}
-
-        impl TaskConfig for TestSchedulableConfig {
-            fn queue_name() -> &'static TaskQueueName {
-                &TEST_QUEUE_NAME
-            }
-            fn max_time_since_last_heartbeat() -> chrono::Duration {
-                chrono::Duration::seconds(60)
-            }
-            // `check_schedule_eligibility` uses the trait default (always Ok)
-            // so the lifecycle test isn't sensitive to property setup.
-        }
-
-        /// User-schedulable queue whose `check_schedule_eligibility` rejects
-        /// any table that carries the `REJECTION_MARKER_PROPERTY`. Used to
-        /// pin the "endpoint surfaces the eligibility-fn 400 verbatim"
-        /// contract that the registry-level unit test cannot exercise
-        /// end-to-end.
-        #[derive(Clone, Debug, Default, Serialize, Deserialize)]
-        #[cfg_attr(feature = "open-api", derive(utoipa::ToSchema))]
-        struct RejectingSchedulableConfig {}
-
-        impl TaskConfig for RejectingSchedulableConfig {
-            fn queue_name() -> &'static TaskQueueName {
-                &REJECTING_QUEUE_NAME
-            }
-            fn max_time_since_last_heartbeat() -> chrono::Duration {
-                chrono::Duration::seconds(60)
-            }
-            fn check_schedule_eligibility(
-                _config: &Self,
-                entity_properties: &std::collections::HashMap<String, String>,
-                _entity: WarehouseTaskEntityId,
-            ) -> Result<(), iceberg_ext::catalog::rest::ErrorModel> {
-                if entity_properties
-                    .get(REJECTION_MARKER_PROPERTY)
-                    .map(String::as_str)
-                    == Some("reject")
-                {
-                    return Err(iceberg_ext::catalog::rest::ErrorModel::bad_request(
-                        format!(
-                            "rejected by test eligibility fn: {REJECTION_MARKER_PROPERTY}=reject"
-                        ),
-                        "RejectedByTestEligibility",
-                        None,
-                    ));
-                }
-                Ok(())
-            }
-        }
-
-        fn build_schema() -> Schema {
-            use iceberg::spec::{NestedField, PrimitiveType};
-            Schema::builder()
-                .with_fields(vec![
-                    NestedField::required(
-                        1,
-                        "id",
-                        iceberg::spec::Type::Primitive(PrimitiveType::Int),
-                    )
-                    .into(),
-                ])
-                .build()
-                .unwrap()
-        }
-
-        #[sqlx::test]
-        async fn schedule_then_409_then_runnow(pool: PgPool) {
-            let (ctx, warehouse, registry) = setup_with_registry(
-                pool,
-                memory_io_profile(),
-                None,
-                AllowAllAuthorizer::default(),
-                TabularDeleteProfile::Hard {},
-                None,
-                1,
-                None,
-            )
-            .await;
-
-            registry
-                .register_queue::<TestSchedulableConfig, TestSchedulablePayload>(
-                    QueueRegistration {
-                        queue_name: &TEST_QUEUE_NAME,
-                        worker_fn: Arc::new(|_| Box::pin(async {})),
-                        num_workers: 0,
-                        scope: QueueScope::Warehouse,
-                        #[cfg(feature = "open-api")]
-                        user_scheduling: UserScheduling::Enabled {
-                            payload_schema: None,
-                        },
-                        #[cfg(not(feature = "open-api"))]
-                        user_scheduling: UserScheduling::Enabled,
-                    },
-                )
-                .await;
-
-            // ---- create namespace + table ----
-            let warehouse_id = warehouse.warehouse_id;
-            let ns = crate::server::test::create_ns(
-                ctx.clone(),
-                warehouse_id.to_string(),
-                "ns1".to_string(),
-            )
-            .await;
-            let ns_params = NamespaceParameters {
-                prefix: Some(Prefix(warehouse_id.to_string())),
-                namespace: ns.namespace.clone(),
-            };
-            let table = CatalogServer::create_table(
-                ns_params,
-                CreateTableRequest {
-                    name: "tab-1".to_string(),
-                    location: None,
-                    schema: build_schema(),
-                    partition_spec: Some(UnboundPartitionSpec::builder().build()),
-                    write_order: None,
-                    stage_create: Some(false),
-                    properties: None,
-                },
-                DataAccess {
-                    vended_credentials: false,
-                    remote_signing: false,
-                },
-                ctx.clone(),
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .unwrap();
-            let table_id = table.metadata.uuid();
-
-            // ---- 1) First schedule call: 200 + fresh task_id ----
-            let resp: ScheduleTaskResponse = ApiServer::schedule_task(
-                warehouse_id,
-                &TEST_QUEUE_NAME,
-                ScheduleTaskRequest {
-                    entity: WarehouseTaskEntityId::Table {
-                        table_id: table_id.into(),
-                    },
-                    scheduled_for: None,
-                    payload: None,
-                },
-                ctx.clone(),
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect("first schedule call should succeed");
-            let first_task_id = resp.task_id;
-
-            // ---- 2) Second schedule call: 409 with the first task_id in body ----
-            let err = ApiServer::schedule_task(
-                warehouse_id,
-                &TEST_QUEUE_NAME,
-                ScheduleTaskRequest {
-                    entity: WarehouseTaskEntityId::Table {
-                        table_id: table_id.into(),
-                    },
-                    scheduled_for: None,
-                    payload: None,
-                },
-                ctx.clone(),
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect_err("second schedule call must hit the unique index");
-            assert_eq!(err.error.code, 409, "expected 409 Conflict, got {err:?}");
-            assert_eq!(err.error.r#type, "TaskAlreadyActive");
-            let id_str = first_task_id.to_string();
-            assert!(
-                err.error.message.contains(&id_str),
-                "409 body must include the existing task-id ({id_str}); got: {}",
-                err.error.message
-            );
-
-            // ---- 3) RunNow on the existing id: succeeds, no error ----
-            ApiServer::control_tasks(
-                warehouse_id,
-                ControlTasksRequest {
-                    action: ControlTaskAction::RunNow,
-                    task_ids: vec![first_task_id],
-                },
-                ctx,
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect("control_tasks RunNow on the existing task should succeed");
-        }
-
-        /// End-to-end pin for the eligibility gate: when a queue's
-        /// `check_schedule_eligibility` returns `Err`, the schedule
-        /// endpoint must surface the same error verbatim — not the
-        /// 409 conflict path, not a generic 500, not a swallowed 200.
-        /// The registry-level dispatch test
-        /// (`test_check_schedule_eligibility_dispatches_through_registry`)
-        /// covers the type-erased closure in isolation; this test covers
-        /// the path from `Service::schedule_task` through the catalog
-        /// fetch into the eligibility fn and back out to the caller.
-        #[sqlx::test]
-        async fn schedule_eligibility_rejection_surfaces_as_400(pool: PgPool) {
-            let (ctx, warehouse, registry) = setup_with_registry(
-                pool,
-                memory_io_profile(),
-                None,
-                AllowAllAuthorizer::default(),
-                TabularDeleteProfile::Hard {},
-                None,
-                1,
-                None,
-            )
-            .await;
-
-            registry
-                .register_queue::<RejectingSchedulableConfig, TestSchedulablePayload>(
-                    QueueRegistration {
-                        queue_name: &REJECTING_QUEUE_NAME,
-                        worker_fn: Arc::new(|_| Box::pin(async {})),
-                        num_workers: 0,
-                        scope: QueueScope::Warehouse,
-                        #[cfg(feature = "open-api")]
-                        user_scheduling: UserScheduling::Enabled {
-                            payload_schema: None,
-                        },
-                        #[cfg(not(feature = "open-api"))]
-                        user_scheduling: UserScheduling::Enabled,
-                    },
-                )
-                .await;
-
-            let warehouse_id = warehouse.warehouse_id;
-            let ns = crate::server::test::create_ns(
-                ctx.clone(),
-                warehouse_id.to_string(),
-                "ns1".to_string(),
-            )
-            .await;
-            let ns_params = NamespaceParameters {
-                prefix: Some(Prefix(warehouse_id.to_string())),
-                namespace: ns.namespace.clone(),
-            };
-            let table = CatalogServer::create_table(
-                ns_params,
-                CreateTableRequest {
-                    name: "tab-reject".to_string(),
-                    location: None,
-                    schema: build_schema(),
-                    partition_spec: Some(UnboundPartitionSpec::builder().build()),
-                    write_order: None,
-                    stage_create: Some(false),
-                    properties: Some(std::collections::HashMap::from([(
-                        REJECTION_MARKER_PROPERTY.to_string(),
-                        "reject".to_string(),
-                    )])),
-                },
-                DataAccess {
-                    vended_credentials: false,
-                    remote_signing: false,
-                },
-                ctx.clone(),
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect("create_table should succeed");
-            let table_id = table.metadata.uuid();
-
-            let err = ApiServer::schedule_task(
-                warehouse_id,
-                &REJECTING_QUEUE_NAME,
-                ScheduleTaskRequest {
-                    entity: WarehouseTaskEntityId::Table {
-                        table_id: table_id.into(),
-                    },
-                    scheduled_for: None,
-                    payload: None,
-                },
-                ctx,
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect_err("eligibility rejection must surface as an error from the endpoint");
-
-            assert_eq!(err.error.code, 400, "expected 400, got {err:?}");
-            assert_eq!(
-                err.error.r#type, "RejectedByTestEligibility",
-                "endpoint must surface the queue's error code verbatim, got {err:?}"
-            );
-            assert!(
-                err.error.message.contains(REJECTION_MARKER_PROPERTY),
-                "endpoint must surface the queue's error message verbatim; got: {}",
-                err.error.message
-            );
-        }
-
-        /// Hitting `schedule` with a queue name that was never registered must
-        /// return `404 QueueNotFound`, not a 400 or a 500. Covers the
-        /// `resolve_schedulable_queue` `None` arm. A real table is required
-        /// because authz runs before queue resolution and would otherwise
-        /// preempt with a 404 for the entity.
-        #[sqlx::test]
-        async fn schedule_unknown_queue_returns_404(pool: PgPool) {
-            let (ctx, warehouse, _registry) = setup_with_registry(
-                pool,
-                memory_io_profile(),
-                None,
-                AllowAllAuthorizer::default(),
-                TabularDeleteProfile::Hard {},
-                None,
-                1,
-                None,
-            )
-            .await;
-            let warehouse_id = warehouse.warehouse_id;
-            let ns = crate::server::test::create_ns(
-                ctx.clone(),
-                warehouse_id.to_string(),
-                "ns1".to_string(),
-            )
-            .await;
-            let table = CatalogServer::create_table(
-                NamespaceParameters {
-                    prefix: Some(Prefix(warehouse_id.to_string())),
-                    namespace: ns.namespace.clone(),
-                },
-                CreateTableRequest {
-                    name: "t-unknown-queue".to_string(),
-                    location: None,
-                    schema: build_schema(),
-                    partition_spec: Some(UnboundPartitionSpec::builder().build()),
-                    write_order: None,
-                    stage_create: Some(false),
-                    properties: None,
-                },
-                DataAccess {
-                    vended_credentials: false,
-                    remote_signing: false,
-                },
-                ctx.clone(),
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect("create_table should succeed");
-
-            let unknown = TaskQueueName::from("never-registered-queue");
-            let err = ApiServer::schedule_task(
-                warehouse_id,
-                &unknown,
-                ScheduleTaskRequest {
-                    entity: WarehouseTaskEntityId::Table {
-                        table_id: table.metadata.uuid().into(),
-                    },
-                    scheduled_for: None,
-                    payload: None,
-                },
-                ctx,
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect_err("unknown queue must not return 2xx");
-            assert_eq!(err.error.code, 404, "expected 404, got {err:?}");
-            assert_eq!(err.error.r#type, "QueueNotFound");
-        }
-
-        /// `tabular_purge` is registered but `user_scheduling: Disabled`.
-        /// Hitting `schedule` against it must return `400` — the destructive
-        /// purge queue must never be exposed via the schedule endpoint, even
-        /// when its name is valid.
-        #[sqlx::test]
-        async fn schedule_non_user_schedulable_queue_returns_400(pool: PgPool) {
-            use crate::service::tasks::tabular_purge_queue::QUEUE_NAME as PURGE_QUEUE_NAME;
-
-            let (ctx, warehouse, _registry) = setup_with_registry(
-                pool,
-                memory_io_profile(),
-                None,
-                AllowAllAuthorizer::default(),
-                TabularDeleteProfile::Hard {},
-                None,
-                1,
-                None,
-            )
-            .await;
-            let warehouse_id = warehouse.warehouse_id;
-            let ns = crate::server::test::create_ns(
-                ctx.clone(),
-                warehouse_id.to_string(),
-                "ns1".to_string(),
-            )
-            .await;
-            let table = CatalogServer::create_table(
-                NamespaceParameters {
-                    prefix: Some(Prefix(warehouse_id.to_string())),
-                    namespace: ns.namespace.clone(),
-                },
-                CreateTableRequest {
-                    name: "t-non-schedulable-queue".to_string(),
-                    location: None,
-                    schema: build_schema(),
-                    partition_spec: Some(UnboundPartitionSpec::builder().build()),
-                    write_order: None,
-                    stage_create: Some(false),
-                    properties: None,
-                },
-                DataAccess {
-                    vended_credentials: false,
-                    remote_signing: false,
-                },
-                ctx.clone(),
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect("create_table should succeed");
-
-            let err = ApiServer::schedule_task(
-                warehouse_id,
-                &PURGE_QUEUE_NAME,
-                ScheduleTaskRequest {
-                    entity: WarehouseTaskEntityId::Table {
-                        table_id: table.metadata.uuid().into(),
-                    },
-                    scheduled_for: None,
-                    payload: None,
-                },
-                ctx,
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect_err("non-schedulable queue must not return 2xx");
-            assert_eq!(err.error.code, 400, "expected 400, got {err:?}");
-            assert_eq!(err.error.r#type, "QueueNotUserSchedulable");
-        }
-
-        static TYPED_PAYLOAD_QUEUE_NAME: LazyLock<TaskQueueName> =
-            LazyLock::new(|| "test_schedulable_typed_payload".into());
-
-        /// Payload type with a required field. A request that omits the field
-        /// (or passes a wrong-shape JSON) fails `serde_json::from_value::<D>`
-        /// in the registry's payload validator, which the endpoint surfaces
-        /// as `400 InvalidTaskPayload`.
-        #[derive(Clone, Debug, Serialize, Deserialize)]
-        struct RequiredFieldPayload {
-            #[allow(dead_code)]
-            must_have: String,
-        }
-        impl TaskData for RequiredFieldPayload {}
-
-        #[derive(Clone, Debug, Default, Serialize, Deserialize)]
-        #[cfg_attr(feature = "open-api", derive(utoipa::ToSchema))]
-        struct TypedPayloadConfig {}
-
-        impl TaskConfig for TypedPayloadConfig {
-            fn queue_name() -> &'static TaskQueueName {
-                &TYPED_PAYLOAD_QUEUE_NAME
-            }
-            fn max_time_since_last_heartbeat() -> chrono::Duration {
-                chrono::Duration::seconds(60)
-            }
-        }
-
-        /// End-to-end pin for the payload validator: a wrong-shape JSON
-        /// payload must surface as `400 InvalidTaskPayload` from the
-        /// endpoint, not a 500 or a silently accepted task. The registry-
-        /// level unit test
-        /// (`test_payload_validator_dispatches_through_registry`) covers
-        /// the type-erased closure in isolation; this test covers the path
-        /// from `Service::schedule_task` through `validate_and_default_payload`
-        /// and back out to the caller.
-        #[sqlx::test]
-        async fn schedule_invalid_payload_returns_400(pool: PgPool) {
-            let (ctx, warehouse, registry) = setup_with_registry(
-                pool,
-                memory_io_profile(),
-                None,
-                AllowAllAuthorizer::default(),
-                TabularDeleteProfile::Hard {},
-                None,
-                1,
-                None,
-            )
-            .await;
-
-            registry
-                .register_queue::<TypedPayloadConfig, RequiredFieldPayload>(QueueRegistration {
-                    queue_name: &TYPED_PAYLOAD_QUEUE_NAME,
-                    worker_fn: Arc::new(|_| Box::pin(async {})),
-                    num_workers: 0,
-                    scope: QueueScope::Warehouse,
-                    #[cfg(feature = "open-api")]
-                    user_scheduling: UserScheduling::Enabled {
-                        payload_schema: None,
-                    },
-                    #[cfg(not(feature = "open-api"))]
-                    user_scheduling: UserScheduling::Enabled,
-                })
-                .await;
-
-            let warehouse_id = warehouse.warehouse_id;
-            let ns = crate::server::test::create_ns(
-                ctx.clone(),
-                warehouse_id.to_string(),
-                "ns1".to_string(),
-            )
-            .await;
-            let table = CatalogServer::create_table(
-                NamespaceParameters {
-                    prefix: Some(Prefix(warehouse_id.to_string())),
-                    namespace: ns.namespace.clone(),
-                },
-                CreateTableRequest {
-                    name: "t-bad-payload".to_string(),
-                    location: None,
-                    schema: build_schema(),
-                    partition_spec: Some(UnboundPartitionSpec::builder().build()),
-                    write_order: None,
-                    stage_create: Some(false),
-                    properties: None,
-                },
-                DataAccess {
-                    vended_credentials: false,
-                    remote_signing: false,
-                },
-                ctx.clone(),
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect("create_table should succeed");
-
-            let err = ApiServer::schedule_task(
-                warehouse_id,
-                &TYPED_PAYLOAD_QUEUE_NAME,
-                ScheduleTaskRequest {
-                    entity: WarehouseTaskEntityId::Table {
-                        table_id: table.metadata.uuid().into(),
-                    },
-                    scheduled_for: None,
-                    // Wrong shape: queue expects `{ must_have: String }`,
-                    // we send a key it doesn't know about.
-                    payload: Some(serde_json::json!({"unexpected": 42})),
-                },
-                ctx,
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect_err("malformed payload must not return 2xx");
-            assert_eq!(err.error.code, 400, "expected 400, got {err:?}");
-            assert_eq!(err.error.r#type, "InvalidTaskPayload");
-        }
-
-        /// Scheduling against a non-existent `table_id` must return `404`,
-        /// not a 500 or a panic. The authz/entity-resolution step short-
-        /// circuits via `AuthZCannotSeeTable::new_not_found` before queue
-        /// resolution gets a chance to run.
-        #[sqlx::test]
-        async fn schedule_missing_table_returns_404(pool: PgPool) {
-            let (ctx, warehouse, registry) = setup_with_registry(
-                pool,
-                memory_io_profile(),
-                None,
-                AllowAllAuthorizer::default(),
-                TabularDeleteProfile::Hard {},
-                None,
-                1,
-                None,
-            )
-            .await;
-
-            // Register a schedulable queue so the failure isn't masked by
-            // queue-resolution; the table miss must still short-circuit
-            // before queue resolution runs.
-            registry
-                .register_queue::<TestSchedulableConfig, TestSchedulablePayload>(
-                    QueueRegistration {
-                        queue_name: &TEST_QUEUE_NAME,
-                        worker_fn: Arc::new(|_| Box::pin(async {})),
-                        num_workers: 0,
-                        scope: QueueScope::Warehouse,
-                        #[cfg(feature = "open-api")]
-                        user_scheduling: UserScheduling::Enabled {
-                            payload_schema: None,
-                        },
-                        #[cfg(not(feature = "open-api"))]
-                        user_scheduling: UserScheduling::Enabled,
-                    },
-                )
-                .await;
-
-            let err = ApiServer::schedule_task(
-                warehouse.warehouse_id,
-                &TEST_QUEUE_NAME,
-                ScheduleTaskRequest {
-                    entity: WarehouseTaskEntityId::Table {
-                        table_id: TableId::new_random(),
-                    },
-                    scheduled_for: None,
-                    payload: None,
-                },
-                ctx,
-                RequestMetadata::new_unauthenticated(),
-            )
-            .await
-            .expect_err("missing table must not return 2xx");
-            assert_eq!(err.error.code, 404, "expected 404, got {err:?}");
-        }
-    }
 }
